Submission/admin_service/requirements.txt
fastapi
psycopg2-binary
pytest
python-dotenv
requests
SQLAlchemy
uvicorn




